Attempt to solve test issues on GitHub actions using interface
